Выделяет логику места назначения в отдельный презентер
